requirements.txt
langchain==0.3.27
langchain-community==0.3.27
langchain-openai==0.3.28
langchain-pinecone==0.2.11
openai==1.99.9
Flask==3.1.3
pypdf==6.1.1
python-dotenv==1.1.1
tiktoken
-e .